Pull a searchDocumentsByName helper into lib
